fix: widen lcd_orientation to uint16_t
Previously stored as uint8_t, which truncated 270 to 14 on assignment.
set_display_defaults() then saw 14 as invalid and reset the field to 0,
so coolerdash sent "orientation": 0 to CoolerControl on every refresh.

Only 270 is affected; 90 and 180 fit in uint8_t.

No other call sites touch the field; CoolerControl JSON layer auto-promotes.

+++ src/device/config.h
@@ -95,7 +95,7 @@ typedef struct Config
     uint16_t display_height;
     float display_refresh_interval;
     uint8_t lcd_brightness;
-    uint8_t lcd_orientation;
+    uint16_t lcd_orientation; // 0/90/180/270
     char display_mode[16];
     char display_background_image_fit[16];
     uint16_t circle_switch_interval;
